+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - The network monitor now marks the chain as unhealthy if it fails to create new blocks ([#1512](https://github.com/0xMiden/miden-node/pull/1512)).
 - Block producer now detects if it is desync'd from the store's chain tip and aborts ([#1520](https://github.com/0xMiden/miden-node/pull/1520)).
 - Pin tool versions in CI ([#1523](https://github.com/0xMiden/miden-node/pull/1523)).
+- Dynamic response variant based on `GetAccount` storage map size ([#1519](https://github.com/0xMiden/miden-node/pull/1519)).
 
 ### Changes